src/requirements.txt
Adafruit-PlatformDetect==3.4.1
adafruit-python-shell==1.3.1
args==0.1.0
asn1crypto==0.24.0
bottle
certifi==2018.8.24
cffi==1.14.5
chardet==3.0.4
clint==0.5.1
cryptography
cycler==0.10.0
docker==3.4.1
docker-pycreds==0.3.0
entrypoints==0.3
future==0.18.2
Glances==3.1.0
idna==2.6
influxdb==5.2.0
iso8601==0.1.14
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.0.1
matplotlib==3.0.2
netifaces==0.10.4
numpy==1.20.1
olefile==0.46
paho-mqtt==1.5.1
Pillow==5.4.1
ply==3.11
psutil
pyasn1==0.4.2
PyAudio==0.2.11
pycparser==2.20
pycrypto
pycryptodomex==3.6.1
PyGObject==3.30.4
pyparsing==2.2.0
pyserial==3.5
pysmi==0.3.2
pysnmp==4.4.6
pystache==0.5.4
python-apt==1.8.4.3
python-dateutil==2.7.3
pytz==2019.1
pyxdg
PyYAML==5.4.1
requests==2.21.0
Rx==3.1.1
scipy==1.6.2
SecretStorage==2.3.1
serial==0.0.97
six==1.12.0
sounddevice==0.4.1
SoundFile==0.10.3.post1
ssh-import-id==5.7
urllib3
websocket-client==0.53.0
